Remove excessive differences over the original RD Web from Windows Server 2012.
The original RD Web pages from Windows Server 2012 R2 were
committed in 80da1889a3d5356e64d2f96242a93a856ed8bb7e.

However, the code in subsequent commit (59d9b4b1352dbc8c) to add OTP was
actually based on the RD Web Pages from Windows Server 2012 (non-R2). This can
be most clearly seen by the fact that the tsportalwebpart.dll is in fact
downgraded by that commit, from 6.3.9600.16384 to 6.2.9200.16384.

In order to more easily see the changes this OTP-version implements, and
also to facility adding in the R2-changes, this commit removes some commented
code, and also fixes up a bunch of immaterial whitespace changes etc compared
to what is available in Server 2012.

Furthermore:
    cook.aspx* is removed because it it unused and appears to be some sort
    of test page.

    webscripts-domain.js was actually introduced by 2012R2 and was just
    not affected when the OTP version based on non-R2 was commited. Remove it
    for now.

    Adding back TSPortalWebPart.resources.dll, which was removed at one point
    without explanation.

With this commit, the state of the Pages directory should be the Server 2012
version plus a fairly clean set of changes to add OTP support. Also, it
introduces the GetRealRequestUri() method, which seems unrelated to OTP but
allows running the site on a port number different from the one that is
publicly used/exposed.