Commit the unchanged RD Web Pages from Windows Server 2012.
Sort of a vendor-branch thing. Note that because this commit is based off
the current state of RDSFactor, which is based on the RD Web Pages from
Windows Server 2012, this commit only shows removal of the modifications done
in this project. But the state of the Pages folder after this commit matches
the full unmodified version from Server 2012.