web: Fix possibility to circumvent 2FA if the application pool is restarted.
Tests show that it's possible to skip 2FA authentication if the application
pool is restarted while the user is currently in tokenform.aspx, and the
user then skips the form and manually try to access Default.aspx.

This is because Default.aspx relies on a session variable to tell it both
that a) 2FA is required and b) 2FA is not yet completed. If the application is
restarted this flag disappears, and so Default.aspx agrees to render its
content.

Fix this by centralizing the 2FA state management, and by always checking the
configuration (not the session) to determine if 2FA is required. The session
variable now only indicates if 2FA have been completed or not.